+++ appfinanzaspersonales/lib/Configuracion.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class ConfiPerfilPantalla extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Configuracion del perfil"),
+      ),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pop(context);
+          },
+          child: Text('Regresar!'),
+        ),
+      ),
+    );
+  }
+}
