+++ appfinanzaspersonales/lib/Gastos.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class GastosPantalla extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Gastos"),
+      ),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pop(context);
+          },
+          child: Text('Regresar!'),
+        ),
+      ),
+    );
+  }
+}
