+++ appfinanzaspersonales/lib/Presupuesto.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class PresupuestosPantalla extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Presupuestos"),
+      ),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pop(context);
+          },
+          child: Text('Regresar!'),
+        ),
+      ),
+    );
+  }
+}
