+++ appfinanzaspersonales/lib/RegistroIngrYEgre.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class RegIngYEgrePantalla extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Registro de Ingresos y Egresos"),
+      ),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pop(context);
+          },
+          child: Text('Regresar!'),
+        ),
+      ),
+    );
+  }
+}
